코드 리뷰 반영 - UserList 폴더 구조 변경, moveMyProfile 함수 리렌더링 방지
